Move models to its own file
