+++ Sources/XcodesKit/Models.swift
@@ -95,119 +95,3 @@ public struct VersionPlist: Decodable {
         case productBuildVersion = "ProductBuildVersion"
     }
 }
-
-struct DownloadableRuntimesResponse: Decodable {
-    let sdkToSimulatorMappings: [SDKToSimulatorMapping]
-    let sdkToSeedMappings: [SDKToSeedMapping]
-    let refreshInterval: Int
-    let downloadables: [DownloadableRuntime]
-    let version: String
-}
-
-public struct DownloadableRuntime: Decodable {
-    let category: Category
-    let simulatorVersion: SimulatorVersion
-    let source: String
-    let dictionaryVersion: Int
-    let contentType: ContentType
-    let platform: Platform
-    let identifier: String
-    let version: String
-    let fileSize: Int
-    let hostRequirements: HostRequirements?
-    let name: String
-    let authentication: Authentication?
-}
-
-struct SDKToSeedMapping: Decodable {
-    let buildUpdate: String
-    let platform: DownloadableRuntime.Platform
-    let seedNumber: Int
-}
-
-extension DownloadableRuntime {
-    struct SimulatorVersion: Decodable {
-        let buildUpdate: String
-        let version: String
-    }
-
-    struct HostRequirements: Decodable {
-        let maxHostVersion: String?
-        let excludedHostArchitectures: [String]?
-        let minHostVersion: String?
-        let minXcodeVersion: String?
-    }
-
-    enum Authentication: String, Decodable {
-        case virtual = "virtual"
-    }
-
-    enum Category: String, Decodable {
-        case simulator = "simulator"
-    }
-
-    enum ContentType: String, Decodable {
-        case diskImage = "diskImage"
-        case package = "package"
-    }
-
-    enum Platform: String, Decodable {
-        case tvOS = "com.apple.platform.appletvos"
-        case iOS = "com.apple.platform.iphoneos"
-        case macOS = "com.apple.platform.macosx"
-        case watchOS = "com.apple.platform.watchos"
-
-        var order: Int {
-            switch self {
-                case .iOS: return 1
-                case .macOS: return 2
-                case .watchOS: return 3
-                case .tvOS: return 4
-            }
-        }
-
-        var shortName: String {
-            switch self {
-                case .iOS: return "iOS"
-                case .macOS: return "macOS"
-                case .watchOS: return "watchOS"
-                case .tvOS: return "tvOS"
-            }
-        }
-    }
-}
-
-struct SDKToSimulatorMapping: Decodable {
-    let sdkBuildUpdate: String
-    let simulatorBuildUpdate: String
-    let sdkIdentifier: String
-}
-
-public struct InstalledRuntime: Decodable {
-    let build: String
-    let deletable: Bool
-    let identifier: UUID
-    let kind: Kind
-    let lastUsedAt: Date?
-    let path: String
-    let platformIdentifier: Platform
-    let runtimeBundlePath: String
-    let runtimeIdentifier: String
-    let signatureState: String
-    let state: String
-    let version: String
-    let sizeBytes: Int?
-}
-
-extension InstalledRuntime {
-    enum Kind: String, Decodable {
-        case bundled = "Bundled with Xcode"
-        case legacyDownload = "Legacy Download"
-    }
-
-    enum Platform: String, Decodable {
-        case tvOS = "com.apple.platform.appletvsimulator"
-        case iOS = "com.apple.platform.iphonesimulator"
-        case watchOS = "com.apple.platform.watchsimulator"
-    }
-}
